C4-98 cleanup a lot, update doc
