+++ src/encoded/types/__init__.py
@@ -308,62 +308,3 @@ class AnnotationField(Item):
     })
     def display_title(self, vcf_name):
         return vcf_name
-
-
-@collection(
-    name='variant-samples',
-    properties={
-        'title': 'Variants (sample)',
-        'description': 'List of all variants with sample specific information',
-    })
-class VariantSample(Item):
-    """Class for variant samples."""
-
-    item_type = 'variant_sample'
-    schema = load_schema('encoded:schemas/variant_sample.json')
-    embedded_list = [
-        'variant.*'  # embed all variant fields
-    ]
-
-    @calculated_property(schema={
-        "title": "Display Title",
-        "description": "A calculated title for every object in 4DN",
-        "type": "string"
-    })
-    def display_title(self, request, CALL_INFO, variant=None):
-        variant = get_item_if_you_can(request, variant, 'Variant')
-        if variant:
-            return CALL_INFO + ' ' + variant['display_title']
-        return CALL_INFO
-
-    @calculated_property(schema={
-        "title": "AD_REF",
-        "description": "Reference AD",
-        "type": "integer"
-    })
-    def AD_REF(self, AD):
-        if AD:
-            return int(AD.split(',')[0])
-        return -1
-
-    @calculated_property(schema={
-        "title": "AD_ALT",
-        "description": "Alternate AD",
-        "type": "integer"
-    })
-    def AD_ALT(self, AD):
-        if AD:
-            return int(AD.split(',')[1])
-        return -1
-
-    @calculated_property(schema={
-        "title": "AF",
-        "description": "Allele Frequency",
-        "type": "number"
-    })
-    def AF(self, AD):
-        if AD:
-            ref, alt = AD.split(',')
-            return round(int(alt) / (int(ref) + int(alt)), 3)  # round to 3 digits
-        return 0.0
-
